Develop jlin frontend

+++ dotnet-backend/BattleArena.Db/TriviaGameQuestion.cs
@@ -7,11 +7,11 @@ public class TriviaGameQuestion : AuditBase
 {
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     [Column("id", Order = 0)]
-    public long Id { get; set; }
+    public int Id { get; set; }
 
     [Column("game_id")]
     public long GameId { get; set; }
 
     [Column("question_id")]
-    public long QuestionId { get; set; }
+    public int QuestionId { get; set; }
 }
